+++ tests/testthat/test_benchmark.R
@@ -601,3 +601,15 @@ test_that("benchmark allows that param_values overwrites tune token", {
   expect_error(benchmark(design), "cannot be trained with TuneToken present in hyperparameter")
 })
 
+test_that("warning when mixing predict types", {
+  design = benchmark_grid(
+    tsk("iris"),
+    list(
+      lrn("classif.debug", predict_type = "prob"),
+      lrn("classif.featureless", predict_type = "response")
+    ),
+    rsmp("cv", folds = 3)
+  )
+  expect_warning(benchmark(design), "Multiple predict types detected")
+  expect_warning(benchmark(design, single_predict_type = FALSE), NA)
+})